fix: 투자 유형 예외 처리 및 평균 수익률 정밀도 개선

+++ back/src/main/java/com/baedang/report/leaderboard/dto/TypeAggregate.java
@@ -1,12 +1,14 @@
 package com.baedang.report.leaderboard.dto;
 
+import java.math.BigDecimal;
+
 /**
- * 유형별 집계 프로젝션(최신 as_of 코호트). JPQL 생성자 표현식으로 채운다. 평균 수익률은 JPQL
- * {@code avg}가 Double 로 주므로 그대로 받고, 표시 단계에서 수익률%로 포맷한다.
+ * 유형별 집계 프로젝션(최신 as_of 코호트). 네이티브 쿼리의 {@code avg(numeric)} 결과를
+ * BigDecimal로 받아 정밀도를 유지하고, 표시 단계에서만 반올림한다.
  */
 public record TypeAggregate(
         String typeCode,
         long count,
-        double avgReturnRate
+        BigDecimal avgReturnRate
 ) {
 }
